Made GameLoop.as contructable with fake classes.
Created an interface file for each of the classes in the parameter list
for GameLoop.as, and reasigned the parameters and variables of all classes
involved to use the interfaces so that i cabn substitute in fake classes
during testing.

I feel like the seperate classes for PlayerModel, CollectableModel and
ObstacleModel may be superfluous. Investigate this later.